refactor(frontend): extract VaultRow to components/VaultRow.jsx (6/N)
Moves the memoized vault-list row out of App.jsx. ~120 LOC of
self-contained UI. memo() wrapper preserved on export. Behavior
identical.

Build clean. App.jsx down to 3452 LOC.
